# Avoid the repeated computation of positions in the barycentric reference frame