--- a/physics/barycentric_rotating_reference_frame.hpp
+++ b/physics/barycentric_rotating_reference_frame.hpp
@@ -96,6 +96,8 @@ class BarycentricRotatingReferenceFrame
 
  private:
   using Base = RigidReferenceFrame<InertialFrame, ThisFrame>;
+  using BodiesToPositions =
+      typename Ephemeris<InertialFrame>::BodiesToPositions;
 
   template<typename SF, typename SB, int o = 0>
   using Trihedron = typename Base::template Trihedron<SF, SB, o>;
@@ -108,11 +110,20 @@ class BarycentricRotatingReferenceFrame
                                     Instant() + NaN<Time>};
   };
 
+  template<int degree>
+  Derivative<Position<InertialFrame>, Instant, degree> PrimaryDerivative(
+      BodiesToPositions const& bodies_to_positions,
+      Instant const& t) const;
+  template<int degree>
+  Derivative<Position<InertialFrame>, Instant, degree> SecondaryDerivative(
+      BodiesToPositions const& bodies_to_positions,
+      Instant const& t) const;
   template<
       int degree,
       std::vector<not_null<MassiveBody const*>> const
           BarycentricRotatingReferenceFrame<InertialFrame, ThisFrame>::*bodies>
   Derivative<Position<InertialFrame>, Instant, degree> BarycentreDerivative(
+      BodiesToPositions const& bodies_to_positions,
       Instant const& t,
       CachedDerivatives& cache) const;