For the rotating-pulsating frame it can result in a 2x speedup if there are many bodies in the frame.
Before:
```
Run on (48 X 3793 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
------------------------------------------------------------------------------------------
Benchmark                                                Time             CPU   Iterations
------------------------------------------------------------------------------------------
BM_BodyCentredNonRotatingReferenceFrame/1024001        279 ms          281 ms            3
BM_BarycentricRotatingReferenceFrame/1024001          3965 ms         3953 ms            1
BM_RotatingPulsatingReferenceFrame/1024001           51332 ms        51297 ms            1
```
After:
```
Run on (48 X 3793 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
------------------------------------------------------------------------------------------
Benchmark                                                Time             CPU   Iterations
------------------------------------------------------------------------------------------
BM_BodyCentredNonRotatingReferenceFrame/1024001        275 ms          281 ms            2
BM_BarycentricRotatingReferenceFrame/1024001          4945 ms         4938 ms            1
BM_RotatingPulsatingReferenceFrame/1024001           24687 ms        24672 ms            1
```